Implement basic quest based commands
Missions in the Mission world can be clicked now but will crash the game if there's no map. If a quest map has been played, the next one will unlock. This also implements quest timers and saving of quest progress to the player's save. This however needs some more work