Dynamic (targets) fixes
There were a couple of issues with the previous approach, since darts reset every monday we should have set the dates to monday of that week and then each one is exactly 7 days after it, the config has first 4 entries as invalid so those had to be removed, so there's 26 valid entries that will wrap around

Dates are now adjusted to Monday at 00:00:00 every week.

There's also a debug level in the function that can be set to 1 or 2 depending on how much debug info is needed to fix any future bugs with this function, for now it's at 0 as it is working as of now.